requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
astor==0.8.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==0.12.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
idna-ssl==1.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
JPype1==0.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
konlpy==0.5.1
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.5.2
nltk==3.4.3
numpy==1.16.3
pandas==0.24.2
protobuf==3.9.1
pyee==6.0.0
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
rouge==0.3.2
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
slackclient==2.1.0
slackeventsapi==2.1.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
yarl==1.3.0
